==> 03_Descriptive.Rmd <==
---
title: "Descriptive Figures"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
setwd("/dados/Analysis/Thiago/Scripts/CoronaVac-Waning/TND")
pacman::p_load(tidyverse,tidylog,tidytable,tableone,gtsummary,rms,patchwork)
source("ggplot_theme_Publication.R")

```


```{r cars}
vacc <- data.table::rbindlist(lapply(Sys.glob("/dados/Analysis/Thiago/DB_processed/db_vacine111121/part-*.parquet"), arrow::read_parquet))
vacc <- vacc %>% mutate(flag_incon=case_when(
  d2_data_aplicacao-d1_data_aplicacao<13~1L,
  d1_data_aplicacao<"2021-01-18"~1L,
  !is.na(d2_nome_vacina )& d2_nome_vacina!=d1_nome_vacina~1L,
  d_idade<12~1L,
  TRUE~0L
))
vacc_clean <- vacc %>% filter(flag_incon==0)

```


# Distribution
```{r}
#main analysis

datebreaks <- seq(as.Date("2021-01-01"), as.Date("2021-11-12"), by = "1 month")
plot_case_control <- vac_test_rt_pcr %>% ggplot(aes(dt_coleta, fill = confirmado)) +
  geom_histogram(binwidth = 7, color = "white", alpha = 0.7, size = 0.25) +
  scale_fill_manual(values = c("Confirmado" = "firebrick2", "Negativo" = "deepskyblue"), labels = c("Cases", "Controls")) +
  labs(y = "Suspected cases by week", x = "Sample collection date", fill = "") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 4e5), breaks = seq(0, 4e5, 1e5)) +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y ", limits = c(as.Date("2021-01-01"), as.Date("2021-11-12"))) +
  theme_Publication() +
  theme(legend.position = "bottom")

plot_case_control
```


# CC
```{r}

plot_cc <- vac_test_rt_pcr %>%
  mutate(vs_type3 = case_when(
    str_detect(vs_type2, "CV_") ~ "CoronaVac",
    vs_type2 == "uv" ~ "Unvaccinated",
    TRUE ~ "Other Vaccines"
  )) %>%
  ggplot(aes(dt_coleta, fill = vs_type3)) +
  geom_histogram(binwidth = 7, color = "white", alpha = 0.7, size = 0.25) +
  labs(y = "Counts", x = "Sample collection date", fill = "") +
  scale_fill_manual(values = c("CoronaVac" = "forestgreen", "Other Vaccines" = "deepskyblue2", "Unvaccinated" = "firebrick2")) +
  scale_y_continuous(labels = scales::comma, limits = c(0, 2e5), breaks = seq(0, 2e5, .5e5)) +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y ", limits = c(as.Date("2021-01-01"), as.Date("2021-11-12"))) +
  theme(legend.position = "bottom") +
  theme_Publication() +
  theme(legend.position = "bottom", panel.spacing.x = unit(2, "lines")) +
  facet_wrap(~confirmado, labeller = as_labeller(c("Negativo" = "Controls", "Confirmado" = "Cases")))

plot_cc
```


# CC - Antigen/RT-PCR
```{r}
plot_cc_ag <- vac_test %>%
  mutate(vs_type3 = case_when(
    str_detect(vs_type2, "CV_") ~ "CoronaVac",
    vs_type2 == "uv" ~ "Unvaccinated",
    TRUE ~ "Other Vaccines"
  )) %>%
  ggplot(aes(dt_coleta, fill = vs_type3)) +
  geom_histogram(binwidth = 7, color = "white", alpha = 0.7, size = 0.25) +
  labs(y = "Counts", x = "Sample collection date", fill = "") +
  scale_fill_manual(values = c("CoronaVac" = "forestgreen", "Other Vaccines" = "deepskyblue2", "Unvaccinated" = "firebrick2")) +
  scale_y_continuous(labels = scales::comma, limits = c(0, 4e5), breaks = seq(0, 4e5, 1e5)) +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y ", limits = c(as.Date("2021-01-01"), as.Date("2021-11-12"))) +
  theme(legend.position = "bottom") +
  theme_Publication() +
  theme(legend.position = "bottom", panel.spacing.x = unit(2, "lines")) +
  facet_wrap(~confirmado, labeller = as_labeller(c("Negativo" = "Controls", "Confirmado" = "Cases")))

plot_cc_ag
```


```{r plot_vaccine_uptake, echo=FALSE}
first_dose_g <- vacc_clean %>% filter(!is.na(d1_data_aplicacao),d1_data_aplicacao>"2021-01-17",d1_data_aplicacao<"2021-11-12")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    )) %>% 
  group_by(d1_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d1_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  ) %>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d1_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent)+ theme_Publication()+labs(x="",y="Cumulative uptake %",color="Age Group", title = "First dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()

second_dose_g <- vacc_clean %>% filter(!is.na(d2_data_aplicacao),d2_data_aplicacao>"2021-02-01",d2_data_aplicacao<"2021-11-12")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    )) %>% 
  group_by(d2_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d2_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  ) %>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d2_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent, limits = c(0,1))+ theme_Publication()+labs(x="",y="",color="Age Group", title = "Second dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()

third_dose_g <- vacc_clean %>% filter(!is.na(d3_data_aplicacao),d3_data_aplicacao>"2021-08-01",d3_data_aplicacao<"2021-11-12")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    ))  %>% 
  group_by(d3_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d3_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  )%>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d3_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent, limits = c(0,1))+ theme_Publication()+labs(x="",y="",color="Age Group", title = "Booster dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()


first_dose_corona <- vacc_clean %>% filter(!is.na(d1_data_aplicacao),d1_data_aplicacao>"2021-01-17",d1_data_aplicacao<"2021-11-12",d1_nome_vacina=="CoronaVac")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    )) %>% 
  group_by(d1_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d1_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  ) %>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d1_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent, limits = c(0,1))+ theme_Publication()+labs(x="",y="Cumulative uptake %",color="Age Group", title = "First dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()

second_dose_corona <- vacc_clean %>% filter(!is.na(d2_data_aplicacao),d2_data_aplicacao>"2021-02-01",d2_data_aplicacao<"2021-11-12",d1_nome_vacina=="CoronaVac")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    )) %>% 
  group_by(d2_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d2_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  ) %>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d2_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent, limits = c(0,1))+ theme_Publication()+labs(x="",y="",color="Age Group", title = "Second dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()

third_dose_corona <- vacc_clean %>% filter(!is.na(d3_data_aplicacao),d3_data_aplicacao>"2021-08-01",d3_data_aplicacao<"2021-11-12",d1_nome_vacina=="CoronaVac")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    ))  %>% 
  group_by(d3_data_aplicacao, faixa_etaria) %>%
  count(name = "vacinacao_por_faixa_etaria") %>%
  group_by(faixa_etaria) %>%
  arrange(d3_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_faixa_etaria)
  )%>% mutate(cobertura_perc_cum=case_when(
  faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60-79"~cobertura_por_dia_cumu/26712827,
  faixa_etaria=="80+"~cobertura_por_dia_cumu/4617408))%>% ggplot(aes(d3_data_aplicacao,cobertura_perc_cum,color=faixa_etaria))+geom_line(size=1, alpha=0.8)+scale_x_date(breaks = "1 month",date_labels = "%B")+scale_y_continuous(labels=scales::percent, limits = c(0,1))+ theme_Publication()+labs(x="",y="",color="Age Group", title = "Booster dose")+theme(axis.text.x = element_text(angle = 90, vjust = 0.5))+ggsci::scale_color_lancet()


first_dose_g+second_dose_g+third_dose_g+first_dose_corona+second_dose_corona+third_dose_corona+ plot_layout(guides = "collect") & theme(legend.position = 'bottom') &plot_annotation(tag_levels = "A")&scale_x_date(limits = c(as.Date("2021-01-01"),as.Date("2021-11-12")),breaks = "1 month",date_labels = "%b/%Y")
second_dose_g+labs(title = "Second dose")

type_graph <- vacc_clean %>% filter(!is.na(d3_data_aplicacao),d3_data_aplicacao>"2021-08-01",d3_data_aplicacao<"2021-11-11",d1_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan")%>% mutate(d1_nome_vacina=case_when(
  d1_nome_vacina=="Covid-19-AstraZeneca"~"AZ",
  d1_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan"~"CV",
  d1_nome_vacina=="Vacina covid-19 - Ad26.COV2.S - Janssen-Cilag"~"Ad26",
  d1_nome_vacina=="Vacina covid-19 - BNT162b2 - BioNTech/Fosun Pharma/Pfizer"~"BNT162b2",
  is.na(d1_nome_vacina)~"uv"),
  d2_nome_vacina=case_when(
    d2_nome_vacina=="Covid-19-AstraZeneca"~"AZ",
    d2_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan"~"CV",
    d2_nome_vacina=="Vacina covid-19 - Ad26.COV2.S - Janssen-Cilag"~"Ad26",
    d2_nome_vacina=="Vacina covid-19 - BNT162b2 - BioNTech/Fosun Pharma/Pfizer"~"BNT162b2"),
  d3_nome_vacina=case_when(
    d3_nome_vacina=="Covid-19-AstraZeneca"~"AZ",
    d3_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan"~"CV",
    d3_nome_vacina=="Vacina covid-19 - Ad26.COV2.S - Janssen-Cilag"~"Ad26",
    d3_nome_vacina=="Vacina covid-19 - BNT162b2 - BioNTech/Fosun Pharma/Pfizer"~"BNT162b2")) %>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    ),comb_3rddose=paste0(d1_nome_vacina,d3_nome_vacina))  %>% 
  group_by(d3_data_aplicacao, comb_3rddose) %>%
  count(name = "vacinacao_por_terceira") %>% 
  group_by(comb_3rddose) %>%
  arrange(d3_data_aplicacao) %>%
  mutate(
    cobertura_por_dia_cumu =
      cumsum(vacinacao_por_terceira)
  )%>% mutate(cobertura_perc_cum=case_when(
  comb_3rddose=="CVAZ"~cobertura_por_dia_cumu/9395171, #Total
  comb_3rddose=="CVCV"~cobertura_por_dia_cumu/9395171,
  comb_3rddose=="CVAd26"~cobertura_por_dia_cumu/9395171,
  comb_3rddose=="CVBNT162b2"~cobertura_por_dia_cumu/9395171),
  comb_3rddose=case_when(comb_3rddose=="CVCV"~"CoronaVac",
                          comb_3rddose=="CVAZ"~"ChAdOx1",
                          comb_3rddose=="CVAd26"~"Ad26.COV2.S",
                          comb_3rddose=="CVBNT162b2"~"BNT162b2"))  %>% complete(nesting(comb_3rddose), d3_data_aplicacao = seq(min(d3_data_aplicacao), as.Date("2021-11-11"), by = "day")) %>%  arrange(d3_data_aplicacao, comb_3rddose) %>%
  fill(cobertura_perc_cum, .direction = "down") 


type_graph %>% ggplot(aes(d3_data_aplicacao, cobertura_perc_cum, fill = factor(comb_3rddose, levels = c("Ad26.COV2.S", "ChAdOx1", "BNT162b2","CoronaVac")))) +
  geom_area(size = 0.2, alpha = 0.8) +
  scale_x_date(breaks = "1 month", date_labels = "%b/%Y") +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  theme_Publication() +
  labs(x = "", y = "% of Booster doses \nby vaccine", fill = "Vaccine:", title = "Fully vaccinated with CoronaVac") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  scale_fill_manual(values = c("Ad26.COV2.S" = "black", "BNT162b2" = "deepskyblue3","ChAdOx1"="orange","CoronaVac" = "forestgreen"))






first_dose_g+second_dose_g+third_dose_g+ plot_layout(guides = "collect") & theme(legend.position = 'bottom') &plot_annotation(tag_levels = "A")&scale_x_date(limits = c(as.Date("2021-01-01"),as.Date("2021-12-12")),breaks = "1 month",date_labels = "%b/%Y")
second_dose_g+labs(title = "Second dose")


first_dose_g+second_dose_g+third_dose_g+first_dose_corona+second_dose_corona+third_dose_corona+ plot_layout(guides = "collect") & theme(legend.position = 'bottom') &plot_annotation(tag_levels = "A")&scale_x_date(limits = c(as.Date("2021-01-01"),as.Date("2021-12-12")),breaks = "1 month",date_labels = "%b/%Y")
second_dose_g+labs(title = "Second dose")



vacc_clean <- vacc %>% filter(flag_incon==0, d_idade>11) %>% mutate(d1_nome_vacina=case_when(
  d1_nome_vacina=="Covid-19-AstraZeneca"~"ChAdOx1",
  d1_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan"~"CoronaVac",
  d1_nome_vacina=="Vacina covid-19 - Ad26.COV2.S - Janssen-Cilag"~"Ad26.COV2.S",
  d1_nome_vacina=="Vacina covid-19 - BNT162b2 - BioNTech/Fosun Pharma/Pfizer"~"BNT162b2"),
  d2_nome_vacina=case_when(
  d2_nome_vacina=="Covid-19-AstraZeneca"~"ChAdOx1",
  d2_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan"~"CoronaVac",
  d2_nome_vacina=="Vacina covid-19 - Ad26.COV2.S - Janssen-Cilag"~"Ad26.COV2.S",
  d2_nome_vacina=="Vacina covid-19 - BNT162b2 - BioNTech/Fosun Pharma/Pfizer"~"BNT162b2"))


cumulatives <- vacc_clean %>% filter(d2_nome_vacina!="Ad26.COV2.S" | is.na(d2_nome_vacina))%>%filter(case_when(
    str_detect(d1_nome_vacina, "Ad26") ~ d1_data_aplicacao > "2021-04-30",
    str_detect(d1_nome_vacina, "BNT") ~ d1_data_aplicacao > "2021-04-30",
    TRUE ~ TRUE
  )) %>% droplevels() %>%  select(d1_data_aplicacao,d2_data_aplicacao,d1_nome_vacina,id_vigvac,d_idade) %>% pivot_longer(-c(id_vigvac,d1_nome_vacina,d_idade), names_to="dose", values_to="data")%>%droplevels() %>%
  mutate(
    faixa_etaria = case_when(
      d_idade < 18 ~ "12-17",
      d_idade < 60 ~ "18-59",
      d_idade >= 60 ~ "60+"
    )
  ) %>% mutate(d1_nome_vacina=fct_relevel(d1_nome_vacina,"BNT162b2","ChAdOx1","CoronaVac","Ad26.COV2.S")) %>%
  group_by(d1_nome_vacina,data,dose, faixa_etaria) %>%
  count(name = "vac_fe_dose") %>%
  group_by(d1_nome_vacina,dose, faixa_etaria) %>%
  arrange(data) %>%  mutate(
    cobertura_por_dia_cumu =
      cumsum(vac_fe_dose)
  )%>%
  mutate(cobertura_perc_cum = case_when(
    faixa_etaria=="12-17"~cobertura_por_dia_cumu/18092894,
  faixa_etaria=="18-59"~cobertura_por_dia_cumu/128714674,
  faixa_etaria=="60+"~cobertura_por_dia_cumu/31330235
  ))

plot_geral <- cumulatives %>%
  filter(!is.na(data)) %>%
  rename(dia = data) %>%
  complete(nesting(d1_nome_vacina, faixa_etaria), dia = seq(min(dia), as.Date("2021-11-11"), by = "day")) %>%
  rename(data = dia) %>%
  arrange(d1_nome_vacina, faixa_etaria, data) %>%
  fill(cobertura_perc_cum, .direction = "down") %>%
  ggplot(aes(data, cobertura_perc_cum, fill = d1_nome_vacina)) +
  geom_area(size = .5, alpha = 0.8, color = "black") +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  theme_Publication() +
  labs(x = "", y = "Cumulative uptake %", fill = "Vaccine", title = "") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5), panel.spacing.x = unit(2, "lines"),panel.spacing.y = unit(2, "lines")) +
  scale_colour_Publication() +
  facet_wrap(~ faixa_etaria + dose, ncol = 2) +
  scale_x_date(date_labels = "%b\n%Y ", limits = c(as.Date("2021-01-01"), as.Date("2021-11-11")), date_breaks = "1 month") +
  theme(legend.position = "bottom", axis.text.x = element_text(angle = 0, vjust = 0.4), )

library(egg)
tag_facet_outside(plot1)

top_labels <- c("First dose","Second dose")
side_labels <- c("12-17","18-59", "60+")
tag_facet_outside(plot_geral, open = c("",""), close=c("",""),
                  tag_fun_top = function(i) top_labels[i],
                  tag_fun_right = function(i) side_labels[i],
                  x=c(0.4,0),y=c(0.5,.5))
vac2 <- tag_facet_outside(plot2,
                  tag_fun_top = function(i) top_labels[i],
                  x=c(0.4,0),y=c(0.5,.5))


plot1+geom_area(size = 0.7, alpha = 0.8)

plot1 <- vacc_clean %>% filter(d2_nome_vacina!="Ad26.COV2.S" | is.na(d2_nome_vacina))%>%filter(case_when(
    str_detect(d1_nome_vacina, "Ad26") ~ d1_data_aplicacao > "2021-04-30",
    str_detect(d1_nome_vacina, "BNT") ~ d1_data_aplicacao > "2021-04-30",
    TRUE ~ TRUE
  )) %>% droplevels() %>%  select(d1_data_aplicacao,d2_data_aplicacao,d1_nome_vacina,id_vigvac,d_idade) %>% pivot_longer(-c(id_vigvac,d1_nome_vacina,d_idade), names_to="dose", values_to="data") %>% droplevels() %>% 
 mutate(d1_nome_vacina=fct_relevel(d1_nome_vacina,"BNT162b2","ChAdOx1","CoronaVac","Ad26.COV2.S")) %>% filter(dose=="d1_data_aplicacao") %>% 
  group_by(d1_nome_vacina,data,dose) %>%
  count(name = "vac_dose") %>% 
  group_by(d1_nome_vacina,dose) %>%
  arrange(data) %>%  mutate(
    cobertura_por_dia_cumu =
      cumsum(vac_dose)
  )%>%
  mutate(cobertura_perc_cum = cobertura_por_dia_cumu/178137803)%>%
  ggplot(aes(data, cobertura_perc_cum, fill = d1_nome_vacina)) +
  geom_area(size = 1, alpha = 0.8) +
  scale_x_date(breaks = "1 month", date_labels = "%B") +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  theme_Publication() +
  labs(x = "Month", y = "Cumulative uptake %", Fill = "Vaccine", title = "") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
  scale_colour_Publication()



plot2 <- vacc_clean %>% filter(d2_nome_vacina!="Ad26.COV2.S" | is.na(d2_nome_vacina))%>%filter(case_when(
    str_detect(d1_nome_vacina, "Ad26") ~ d1_data_aplicacao > "2021-04-30",
    str_detect(d1_nome_vacina, "BNT") ~ d1_data_aplicacao > "2021-04-30",
    TRUE ~ TRUE
  )) %>% droplevels() %>%  select(d1_data_aplicacao,d2_data_aplicacao,d1_nome_vacina,id_vigvac,d_idade) %>% pivot_longer(-c(id_vigvac,d1_nome_vacina,d_idade), names_to="dose", values_to="data") %>% droplevels() %>% 
 mutate(d1_nome_vacina=fct_relevel(d1_nome_vacina,"BNT162b2","ChAdOx1","CoronaVac","Ad26.COV2.S")) %>%  filter(dose=="d2_data_aplicacao") %>% 
  group_by(d1_nome_vacina,data,dose) %>%
  count(name = "vac_dose") %>% 
  group_by(d1_nome_vacina,dose) %>%
  arrange(data) %>%  mutate(
    cobertura_por_dia_cumu =
      cumsum(vac_dose)
  )%>%
  mutate(cobertura_perc_cum = cobertura_por_dia_cumu/178137803)%>%
  ggplot(aes(data, cobertura_perc_cum, fill = d1_nome_vacina)) +
  geom_area(size = 1, alpha = 0.8) +
  scale_x_date(breaks = "1 month", date_labels = "%B") +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  theme_Publication() +
  labs(x = "Month", y = "Cumulative uptake %", Fill = "Vaccine", title = "") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5), legend.position = "none") +
  scale_colour_Publication() + guides(fill="none")



library(egg)

tag_facet_outside(plot1)

top_labels <- c("First dose","Second dose")
side_labels <- c("BNT162b2","ChAdOx1", "CoronaVac")
tag_facet_outside(plot1, 
                  tag_fun_top = function(i) top_labels[i],
                  x=c(0.4,0),y=c(0.5,.5))
vac2 <- tag_facet_outside(plot2,
                  tag_fun_top = function(i) top_labels[i],
                  x=c(0.4,0),y=c(0.5,.5))

vaccination <- plot2+geom_area(color="black", alpha=0.6)+labs(fill="Vaccine")+ggsci::scale_fill_jama()
```



```{r}


death %>%
  filter(srag_dt_evoluca > "2020-01-11", srag_dt_evoluca < "2021-11-12") %>%
  ggplot(aes(srag_dt_evoluca)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  theme_Publication() +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y", limits = c(as.Date("2020-02-01"), as.Date("2021-11-11"))) +
  theme(axis.text.x = element_text(angle = 0, vjust = .25), legend.position = "bottom") +
  labs(x = "Date of death", y = "Number of COVID-19 deaths\n by week") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 1.5e4), breaks = seq(0, 1.5e4, .5e4))


vac_test_rt_pcr %>%
  filter(!duplicated(id_vigvac)) %>%
  filter(d1_nome_vacina == "CV", time_2nd_3rd > 90) %>%
  ggplot(aes(time_2nd_3rd)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  scale_x_continuous(limits = c(0, 300), breaks = seq(0, 300, 50)) +
  labs(x = "Interval between second dose of CoronaVac\nand booster dose (days)", y = "Number of people") +
  theme_Publication()




vac_test_rt_pcr %>%
  filter(!duplicated(id_vigvac)) %>%
  filter(str_detect(vs_type2,"CV_v3"),str_detect(vs_type2,"BNT")) %>%
  ggplot(aes(days_vacc_3_test, fill=idade_cat)) +
  geom_bar(alpha=0.75,size=.25,color="white") +
  scale_y_continuous(limits = c(0, 400), breaks = seq(0, 400, 50)) +
  scale_x_continuous(limits = c(0, 60), breaks = seq(0, 60, 15)) +
  expand_limits(x=0)+
  ggsci::scale_fill_npg(limits = rev(levels(vac_test_rt_pcr$idade_cat)))+
  labs(x = "Time between Booster with BN162b2 and sympton onset (days)", y = "Number of people", fill="Age group") +
  theme_Publication()+guides(fill=guide_legend(reverse = TRUE))

```


```{r}
filename <-"Infection_2911.xlsx"

sheets <- openxlsx::getSheetNames(filename)
SheetList <- lapply(sheets,openxlsx::read.xlsx,xlsxFile=filename)
names(SheetList) <- sheets
infection_df <- map_df(SheetList,  ~as.data.frame(.x), .id="id")


filename <-"Hosp_Death_2911.xlsx"

sheets <- openxlsx::getSheetNames(filename)
SheetList <- lapply(sheets,openxlsx::read.xlsx,xlsxFile=filename)
names(SheetList) <- sheets
hospde_df <- map_df(SheetList,  ~as.data.frame(.x), .id="id")

figure_1 <- bind_rows(infection_df,hospde_df,.id="outcome")
unique(figure_1$term)
figure_1 <- figure_1 %>% filter(!str_detect(id,"sensibilidade"),!str_detect(term,"v1_"))%>% mutate(term=str_remove(term,"vs_type2CV_"),
                                                                           term=str_remove(term,"v1_|v2_|v3_"),
                                                                           term=factor(term,levels = unique(term)),
                                                                           id=fct_relevel(factor(id),"Overall","18-59","60-79","80+"))


figure_1 %>% mutate(term=fct_recode(term,"Booster\n w/ BNT162b2"="BNT162b2",
                                    ">180"="180+"),
                    id=fct_recode(id,
                                  "≥80"="80+")) %>%  ggplot(aes(x=term,y=VE, group=outcome, color=outcome))+geom_point(size=1.5)+geom_errorbar(aes(ymin=Lower,ymax=Upper),width=0, color="black")+geom_hline(aes(yintercept=0.5), linetype=2,size=.3)+facet_wrap(~ id, scales = "free") +
  theme_Publication() + theme(panel.border=element_blank(), axis.line=element_line(),axis.text.x = element_text(angle = 90, vjust=.3), legend.position = "bottom")+scale_y_continuous(labels = scales::percent, breaks =seq(0,1,.2), limits = c(-.1,1))+scale_color_manual(labels = c("Infection", "Hospitalization or Death"), values = c("blue", "red"))+labs(y="Vaccine Effectiveness",x="",color="Outcome")+coord_cartesian(ylim=c(0,1))




```



```{r}
lab_data <- data.table::rbindlist(lapply(Sys.glob("/dados/Analysis/Thiago/DB_processed/tnd_processed111121/part-*.parquet"), arrow::read_parquet))

lab_data <- lab_data %>% mutate(
  dt_coleta = case_when(
    is.na(dt_coleta) & confirmado == "Confirmado" ~ dt_notificacao,
    TRUE ~ dt_coleta
  ),
  dt_inicio_sintomas = if_else(is.na(dt_inicio_sintomas), dt_coleta, dt_inicio_sintomas)
)
lab_data <- lab_data %>% filter(dt_notificacao>"2020-02-24", dt_notificacao+14>=dt_inicio_sintomas)

infection <- lab_data %>% filter(!duplicated(paste0(id_vigvac,dt_coleta,confirmado)))


datebreaks <- seq(as.Date("2020-02-01"), as.Date("2021-11-11"), by="3 month")

infection %>%
  filter(confirmado == "Confirmado") %>%
  ggplot(aes(dt_inicio_sintomas)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  theme_Publication() +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y", limits = c(as.Date("2020-02-01"), as.Date("2021-11-12"))) +
  theme(axis.text.x = element_text(angle = 0, vjust = .25), legend.position = "bottom") +
  labs(x = "Date of symptom onset", y = "SARS-CoV-2 Infection\n by week", fill = "") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 4e5), breaks = seq(0, 4e5, 1e5))

hospitalization <- lab_data %>% arrange(srag_dt_interna) %>%  filter(!duplicated(paste0(id_vigvac,srag_dt_interna,confirmado)))

hospitalization %>%
  filter(!is.na(srag_dt_interna), confirmado == "Confirmado", srag_dt_interna > "2020-01-11", srag_dt_interna < "2021-11-11") %>%
  ggplot(aes(srag_dt_interna)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  theme_Publication() +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y", limits = c(as.Date("2020-02-01"), as.Date("2021-11-12"))) +
  theme(axis.text.x = element_text(angle = 0, vjust = .25), legend.position = "bottom") +
  labs(x = "Date of admission", y = "Number of COVID-19 hospitalizations\n by week") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 4e4), breaks = seq(0, 4e4, 1e4))


death <- lab_data %>% arrange(srag_dt_evoluca) %>%  filter(!duplicated(paste0(id_vigvac,srag_dt_interna,confirmado))) %>% filter(confirmado=="Confirmado" & srag_evolucao %in% c(2,3))


death %>%
  filter(srag_dt_evoluca > "2020-01-11", srag_dt_evoluca < "2021-11-12") %>%
  ggplot(aes(srag_dt_evoluca)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  theme_Publication() +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y", limits = c(as.Date("2020-02-01"), as.Date("2021-11-11"))) +
  theme(axis.text.x = element_text(angle = 0, vjust = .25), legend.position = "bottom") +
  labs(x = "Date of death", y = "Number of COVID-19 deaths\n by week") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 1.5e4), breaks = seq(0, 1.5e4, .5e4))


# Severe outcome

only_death <- anti_join(death,hospitalization,by="id_vigvac")

severe_outcomes <- bind_rows(only_death,hospitalization) %>% select(id_vigvac,confirmado,srag_hospital,srag_dt_interna,srag_evolucao,srag_dt_evoluca) %>% mutate(data_severe=coalesce(srag_dt_interna,srag_dt_evoluca))

severe_plot <- severe_outcomes %>% filter(!is.na(data_severe),confirmado=="Confirmado",data_severe>"2020-02-11", data_severe<"2021-11-11")%>% ggplot(aes(data_severe))+ geom_histogram(binwidth = 7, fill="darkred", alpha=0.7,color="black")+theme_Publication()+scale_x_date(breaks = datebreaks,date_labels = "%b\n%Y ",limits=c(as.Date("2020-02-01"),as.Date("2021-11-11")))+theme(axis.text.x = element_text(angle=0, vjust = .25), legend.position = "bottom")+labs(x="Date of hospitalization or death", y= "Weekly\nhospitalization or death")+scale_y_continuous(labels = scales::comma, limits = c(0,4e4), breaks=seq(0,4e4,1e4))


severe_outcomes %>%
  filter(data_severe > "2020-01-11", srag_dt_evoluca < "2021-11-12") %>%
  ggplot(aes(data_severe)) +
  geom_histogram(binwidth = 7, alpha=0.75,size=.25,color="white") +
  theme_Publication() +
  scale_x_date(breaks = datebreaks, date_labels = "%b\n%Y", limits = c(as.Date("2020-02-01"), as.Date("2021-11-11"))) +
  theme(axis.text.x = element_text(angle = 0, vjust = .25), legend.position = "bottom") +
  labs(x = "Date of hospitalization or death", y = "Number of COVID-19 hospitalizations or deaths\n by week") +
  scale_y_continuous(labels = scales::comma, limits = c(0, 4e4), breaks = seq(0, 4e4, 1e4))


```


```{r}
vacc %>% filter(!is.na(d1_data_aplicacao),d1_data_aplicacao>"2021-01-17",d1_data_aplicacao<"2021-11-12",d1_nome_vacina=="Covid-19-Coronavac-Sinovac/Butantan")%>%mutate(
    faixa_etaria = case_when( 
      d_idade <18~"12-17",
      d_idade <60 ~ "18-59",
      d_idade < 80 ~ "60-79",
      d_idade >= 80 ~ "80+"
    )) %>% group_by(faixa_etaria) %>% count(is.na(d2_nome_vacina))
```


```{r}

teste <- as.Date(voc$periodo, format = "%b %Y")
voc <-readxl::read_excel("Data.xlsx")
voc <- janitor::clean_names(voc)
voc_jul <- voc %>%
  mutate(
    variante = case_when(
      classificacao == "Não VOC/VOI"~ "Non VOC/VOI",
      str_detect(variante, "Gama")~"Gamma",
      str_detect(variante,"Delta")~"Delta",
      TRUE~variante)
  ) %>%
  mutate(variante = fct_lump(variante, 3, w = quantidade)) %>%
  mutate(
    periodo = fct_recode(factor(periodo),
                         "Jan/2021" = "Jan 2021",
                         "Feb/2021" = "Fev 2021",
                         "Mar/2021" = "Mar 2021",
                         "Apr/2021" = "Abr 2021",
                         "May/2021" = "Mai 2021",
                         "Jun/2021" = "Jun 2021",
                         "Jul/2021" = "Jul 2021",
                         "Aug/2021" = "Ago 2021",
                         "Sept/2021" = "Set 2021",
                         "Oct/2021" = "Out 2021",
                         "Nov/2021" = "Nov 2021"
    ),
    periodo = fct_relevel(periodo,"Nov/2020","Dec/2020", "Jan/2021", "Feb/2021", "Mar/2021", "Apr/2021", "May/2021", "Jun/2021", "Jul/2021","Aug/2021","Sept/2021","Oct/2021","Nov/2021"),
    variante=fct_relevel(variante,"Non VOC/VOI","Gamma","Delta","Other")
  ) %>% mutate(regiao = recode_factor(factor(regiao),
                                            "Nordeste" = "Northeast",
                                            "Norte" = "North",
                                            "Sudeste" = "Southeast",
                                            "Sul" = "South",
                                            "Centro-oeste" = "Central West"),
               periodo=lubridate::my(periodo)) 

datebreaks <- seq(as.Date("2021-01-01"), as.Date("2021-11-12"), by="1 month")
voc_jul %>% mutate(periodo=periodo) %>% 
    group_by(periodo,regiao) %>%
    count(variante,wt=quantidade) %>%
    mutate(percentualnogrupo =(n / sum(n))) %>%
    ggplot(aes(periodo, y = percentualnogrupo, fill = variante)) +geom_col()+theme_minimal()+ggsci::scale_fill_lancet()+scale_y_continuous(labels = scales::percent)+
  labs(x="", y="% Variant", fill="Variant")+theme_Publication()+theme(axis.text.x = element_text(angle=90, vjust=.5), legend.position = "bottom")+scale_x_date(breaks = datebreaks,date_labels = "%b/%Y ",limits=c(as.Date("2020-12-10"),as.Date("2021-11-15")))+facet_wrap(~regiao,nrow=1)

```




```{r}

patchwork <- (infection_plot / severe_plot/variant)/plot11/plot12 + plot_layout(guides = 'collect')&   theme(axis.title.y = element_text(size=8),legend.position="bottom")

 

# Remove title from second subplot
patchwork[[1]] = patchwork[[1]] + theme(axis.text.x = element_blank(),
                                        axis.ticks.x = element_blank(),
                                        axis.title.x = element_blank())
patchwork[[2]] = patchwork[[2]] + theme(axis.text.x = element_blank(),
                                        axis.ticks.x = element_blank(),
                                        axis.title.x = element_blank())
patchwork[[3]] = patchwork[[3]] + theme(axis.text.x = element_blank(),
                                        axis.ticks.x = element_blank(),
                                        axis.title.x = element_blank())
patchwork[[4]] = patchwork[[4]] + theme(axis.text.x = element_blank(),
                                        axis.ticks.x = element_blank(),
                                        axis.title.x = element_blank())
patchwork[[5]] = patchwork[[5]] + theme(axis.title.x = element_blank())

# Remove title from third subplot


patchwork
```
